requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic==2.7.1
gtts==2.5.1
pillow==10.3.0
aiofiles==23.2.1
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.127.0
google-cloud-vision==3.7.2
pytesseract==0.3.10
jinja2==3.1.4
python-dotenv==1.0.1
